refactor: collect files for redux in store directory
