+++ src/store/action.ts
@@ -1,5 +1,5 @@
 import {createAction} from '@reduxjs/toolkit';
-import { Offers } from './types/offer';
+import { Offers } from '../types/offer';
 
 export const Action = {
   CHANGE_CITY: 'changeCity',
