+++ src/store/index.ts
@@ -1,4 +1,4 @@
 import {configureStore} from '@reduxjs/toolkit';
-import { reducer } from '../reducer';
+import { reducer } from './reducer';
 
 export const store = configureStore({reducer});
